# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: StreamPro — Enterprise TV Streaming Platform API
servers:
  - url: /api/v1
    description: Versioned API base path
...
paths:
...
  /favorites:
    get:
      operationId: getFavorites
      tags: [favorites]
      summary: Get user favorites
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Favorites list
    post:
      operationId: addFavorite
      tags: [favorites]
      summary: Add to favorites
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [channel, movie, series]
                id:
                  type: string
              required: [type, id]
      responses:
        "201":
          description: Added
...
  /subscriptions/plans:
    get:
      operationId: getPlans
      tags: [subscriptions]
      summary: List subscription plans
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Plan list
    post:
      operationId: createPlan
      tags: [subscriptions]
      summary: Create plan (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanDto"
      responses:
        "201":
          description: Plan created
  /subscriptions/me:
    get:
      operationId: getMySubscription
      tags: [subscriptions]
      summary: Get current user's active subscription
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscription or null
  /subscriptions:
    post:
      operationId: subscribe
      tags: [subscriptions]
      summary: Create subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                planId:
                  type: string
                paymentReference:
                  type: string
                couponCode:
                  type: string
              required: [planId]
      responses:
        "201":
          description: Subscription created
  /subscriptions/verify:
    post:
      operationId: verifyPayment
      tags: [subscriptions]
      summary: Verify payment and activate subscription
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reference:
                  type: string
                gateway:
                  type: string
              required: [reference, gateway]
      responses:
        "200":
          description: Verified and activated
  /subscriptions/coupons:
    post:
      operationId: createCoupon
      tags: [subscriptions]
      summary: Create coupon (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponDto"
      responses:
        "201":
          description: Coupon created
  /subscriptions/coupons/apply:
    post:
      operationId: applyCoupon
      tags: [subscriptions]
      summary: Apply coupon code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                planId:
                  type: string
              required: [code, planId]
      responses:
        "200":
          description: Discounted price returned

  # ── PAYMENTS ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    CreatePlanDto:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        price:
          type: number
        currency:
          type: string
          default: USD
        durationDays:
          type: integer
        features:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required: [name, slug, price, durationDays]
    CreateCouponDto:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        maxUses:
          type: integer
        expiresAt:
          type: string
          format: date-time
      required: [code, discountType, discountValue]
...